Changed score up and down for Web Server
One of my team members informed me of an edge case that could happen when using score up and down when swapping teams that could cause team 2 score up to affect team one and vice versa. This appeared to be linked more towards the State Manager getting confused when swapping players, but can't technically reproduce without the Web Server endpoints being used, so I've updated the calls to directly change the widgets to be more reliable in that aspect.