Create a scoreboard page on the stage strike app.
Concept here is that I run streams for small house events where getting
to the stream PC to update set info is... tedious at best. So, I created
a way to update the current stream info with my phone.

I tried to mimic the style of the existing code as much as possible, I
hope this can get upstreamed without too many issues!

External to the new app page:

- Added react-router-dom to make any request not for /scoreboard fall back to the stage strike app.
- Added a route to get the current scoreboard state.
- Ran the build process, since it seems like that is how this repo handles npm build artifacts.

Features of the new page:
- Display current set
- Edit vital stats of any player from either team.
- Adjust set score
- Load sets a la `TSHSelectSetWindow.py` (includes filtering, and "hide concluded sets" option)
- Reasonably responsive to work well with either a mobile device or bigger screen.
- Set up with i18n tags, although I only filled out the en entries.