Rework the threading within the startgg data provider to not use qt events.
I added a result object to our Worker abstraction so that it can be used in a futures-like way.
This allows us to avoid calling processEvents in our worker threads that was causing crashes. I also did some
slight cleanup around the arguments to the TDP and unified the usage of threadpools. Unifying the threadpools
was made possible since we're awaiting the completion of the workers we create, rather than waiting on the
whole threadpool. Common advice for threadpools in qt in general is to only have one master one, so I felt unifying
the pools for different requests in this specific context was warranted.