Added two decorators to marshal functions onto the main thread always.
Qt is extremely twitchy about any non-gui thread touching the gui; even
for reads. More or less the only thing you are reliably allowed to do
from my understanding is emit signals, which then get processed in the
main thread. This PR re-works the webserver routes with this in mind.
The webserver is low-volume, so pushing these things onto the main thread
is not a significant burden.